testing/web-platform/mozilla/tests/indic-detection/tboomi.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
%NÊY”» (Mars) wÁVfji¶T¦§Å EÔ[ JÃ &L”Ô Bj¶. C© wÁVàÆÃ«© S”ßL”Y© &L”[”L EÔ[©.Cfji¶T¦§Å ·LotÛV &L”[”] ´Râfj Ai¦RR”L CW¡P”Y© tÛV &L”[”L %NÊY”» CÃfg\©. &U]”}~]À Cf&L”Øfj &T”ÀfLPÍÕß %TV+Wo w}~¾Ô[]À. CRß &UÚTW±²Å L”Q±Ti¶ CÃ¶´ Bf+Ni Cf&L”+[o %N«¬\U”Lf L”}ig\©.[12] CR]”&X&V CRÚjo %NÊY”» Gß\ %TVÀ HÚT}P©. JÃ ´ËN”À &L”[”] CRß &UÚTW±´ N«§WàÅ EÔ[©&T”Å g¡Qf jÐL+[¾¶, ´Ë¼Å EÔ[© &T”ß\ GÁU+XLÔ, TÔ[¦R”fjLÔ, T”+XY]lLÔ, Tàº~V ©ÃY± Tj§L+[¾¶ %L”¡P©. %NÊY”¼ß vZÚtfL”X¹¶, TÃY U”Ú\lLØ¶ ´Ëfj EÔ[+R± &T”ß\+Y&V. wÁV U¡PX¦©Ô ·L EVWU”] JÆ¶Tv U+X¾¶, ·L±%TÁV %Nlj¦©± TÔ[¦R”fjLØÔ Jß\”] UÁ]À TÔ[¦R”fj¶ %NÊY”¼&X&V EÔ[].
<script>
test(function() {
  assert_equals(document.characterSet, "windows-1252");
},"Should fall back to windows-1252");
</script>
The text content above is a converted extract from the start of the Wikipedia article <a href="https://ta.wikipedia.org/w/index.php?title=%E0%AE%9A%E0%AF%86%E0%AE%B5%E0%AF%8D%E0%AE%B5%E0%AE%BE%E0%AE%AF%E0%AF%8D_(%E0%AE%95%E0%AF%8B%E0%AE%B3%E0%AF%8D)&oldid=3129711">&#2970;&#3014;&#2997;&#3021;&#2997;&#3006;&#2991;&#3021; (&#2965;&#3019;&#2995;&#3021;)</a> and
is licensed under the <a href="https://creativecommons.org/licenses/by-sa/3.0/legalcode">Creative Commons Attribution-ShareAlike 3.0 Unported</a> license.
